# LMS_CODEBASE/packages/openapi/ulms-api.yaml
openapi: 3.0.3
info:
  title: ULMS API
  version: 1.7.0
  description: |
    Unisoft Loan Management System — public API (source of truth per PLANNING/05).
    Conventions: problem+json errors (stable `code`, registry docs/errors.md),
    Idempotency-Key on creating POSTs (replays return the original response
    with `Idempotent-Replay: true`, 48h window), list responses use the
    {data, meta} envelope, money is integer minor units, bilingual objects
    {en,bn} render ONE per request.
    Mutating routes are role-gated (403 ULMS-FORBIDDEN) — realm roles per 06 §1;
    the approval act() ladder-level check maps roles per the band table.
    v1.2 (P2): assessments (CIB/score/DBR/collateral), disbursements with dual
    authorization, compliance EOD/classification/calculator.
    v1.2.1: CIB file channel + 24-month repayment track + raw kept only in the
    encrypted object store (06 §5) + approver inbox.
    v1.3: optimistic concurrency (PATCH version ⇒ 409 ULMS-STATE-0002),
    collateral valuation history, application detail incl. workflow state,
    ?officer= list filter, CIB retry×5 + RB-05 alert on outage.
    v1.4 (P3): payments + rails (intent→redirect→HMAC webhook idempotent by
    UTR), schedules/statements (JSON+CSV), settle quotes, reschedules,
    collections (worklist/actions/PTP/field tasks), tax certificates,
    borrower portal projections.
    v1.4.1: ESCALATE approval action, dunning-ladder timers (queue/run/
...
    portfolio reporting read model.
...
servers:
...
security: [{bearerAuth: []}]
paths:
...
  /compliance/eod/run:
    post:
      tags: [compliance]
      summary: BRPD 15/2024 EOD batch (operator trigger; also scheduled 23:30) — idempotent per run-date with rerun semantics
      requestBody:
        required: false
        content:
          application/json:
            schema: {type: object, properties: {date: {type: string, format: date}}}
      responses:
        "200": {description: run summary, content: {application/json: {schema: {$ref: "#/components/schemas/EodRun"}}}}
...
  /compliance/provision-calculator:
    get:
      tags: [compliance]
      summary: Provision calculator — the policy oracle shared with the frontend (03)
      parameters: [{name: classification, in: query, required: true, schema: {type: string}}, {name: outstandingMinor, in: query, required: true, schema: {type: integer}}]
      responses:
        "200": {description: provision, content: {application/json: {schema: {$ref: "#/components/schemas/ProvisionCalc"}}}}
        "422": {$ref: "#/components/responses/Problem"}
...
  /compliance/provision-jv:
    post:
      tags: [compliance]
      summary: Post the EOD provision total to the Fineract GL as a zero-sum JV (03 "JV queue to Fineract GL") — debit provision expense, credit loan-loss reserve; idempotent per run-date; a drifted rerun total after posting ⇒ 409 + PROVISION_JV_DRIFT alert
      requestBody:
        required: false
        content:
          application/json:
            schema: {type: object, properties: {date: {type: string, format: date}}}
      responses:
        "200": {description: "zero-sum entry incl. Fineract transaction id"}
        "409": {description: drift after posting — reversing entry is a bank process}
    get:
      tags: [compliance]
      summary: Latest posted provision JVs (zero-sum evidence chain)
      responses:
        "200": {description: entries, content: {application/json: {schema: {type: array, items: {type: object}}}}}
...
  /certificates/tax/{year}/pdf:
    get:
      tags: [servicing]
      summary: "Bilingual tax certificate as PDF (openhtmltopdf render, R10 P-E)"
      parameters:
        - {name: year, in: path, required: true, schema: {type: integer}}
        - {name: cif, in: query, required: true, schema: {type: string}}
      responses:
        "200": {description: "PDF document", content: {application/pdf: {schema: {type: string, format: binary}}}}
  /certificates/tax/{year}:
    get:
      tags: [loans]
      summary: Tax certificate — interest paid per loan for the year (03 mod-servicing)
      parameters: [{name: year, in: path, required: true, schema: {type: integer}}, {name: cif, in: query, required: true, schema: {type: string}}]
      responses:
        "200": {description: list envelope, content: {application/json: {schema: {type: object, properties: {data: {type: array, items: {type: object, properties: {cifNo: {type: string}, nameEn: {type: string}, year: {type: integer}, loanNo: {type: string}, interestPaidMinor: {type: integer}, totalPaidMinor: {type: integer}}}}, meta: {$ref: "#/components/schemas/Meta"}}}}}}

  # ================= R3 (v1.6) — origination completion =================
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  responses:
    Problem:
      description: error
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
  schemas:
    Problem:
      type: object
      properties:
        type: {type: string}
        title: {type: string}
        status: {type: integer}
        code: {type: string, description: "stable code — registry docs/errors.md (ULMS-NOT-FOUND, ULMS-FORBIDDEN, ULMS-VAL-0001, ULMS-STATE-0001)"}
        detail: {type: string}
        requestId: {type: string, format: uuid}
        fields:
          type: array
          items:
            type: object
            properties: {path: {type: string}, code: {type: string}, message: {type: string}, messageBn: {type: string}}
    Meta:
      type: object
      properties:
        requestId: {type: string, format: uuid}
        page: {type: integer}
        size: {type: integer}
        totalElements: {type: integer}
        totalPages: {type: integer}
...
    EodRun:
      type: object
      properties:
        runDate: {type: string, format: date}
        loansClassified: {type: integer}
        totalOutstandingMinor: {type: integer}
        totalProvisionMinor: {type: integer}
        provisionsByClass: {type: object, additionalProperties: {type: integer}}
...
    ProvisionCalc:
      type: object
      properties:
        classification: {type: string}
        provisionRateBp: {type: integer, description: basis points (100 = 1%)}
        interestSuspense: {type: boolean}
        provisionMinor: {type: integer}
...